+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Upcoming
 * Removed "Include layers" constraint and instead added a constraint to export only layers.
 * Removed "Include layer groups" constraint and instead added a constraint to export only layer groups. 
 * Removed "Include empty layer groups" constraint without replacement. Empty folders are no longer created.
+* With the above changes, all constraints now consistently only remove (filter) layers. For brevity, the names of all constraints have been shortened. For example, "Only visible layers" has been renamed to "Visible", "Only layers selected in preview" to "Selected in preview", etc.
 * Layer groups as images and as folders are now displayed separately in the preview.
 * Fixed a crash and inconsistencies in the preview if both "Ignore folder structure" and "Only layers selected in preview" are enabled. This also adversely causes the preview to be slower upon selecting items when "Only layers selected in preview" is enabled. The drop in performance is only noticeable for hundreds of layers in an image.